recipes/requirements.txt
blinker==1.7.0
cachelib==0.13.0
click==8.1.7
colorama==0.4.6
cs50==9.3.4
Flask==3.0.2
Flask-Assets==2.1.0
Flask-Scss==0.5
Flask-Session==0.8.0
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
msgspec==0.18.6
packaging==24.0
pyScss==1.4.0
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
termcolor==2.4.0
typing_extensions==4.11.0
webassets==2.0
Werkzeug==3.0.1
gunicorn
google-cloud-datastore
